Refactor DomainEventCollector in ActivityBundle
