+++ src/Sulu/Bundle/ActivityBundle/Resources/config/services.xml
@@ -24,10 +24,11 @@
             <argument>%sulu_activity.storage.persist_payload%</argument>
         </service>
 
-        <service id="sulu_activity.domain_event_collector_subscriber" class="Sulu\Bundle\ActivityBundle\Infrastructure\Doctrine\Subscriber\DomainEventCollectorSubscriber">
+        <service id="sulu_activity.domain_event_collector_subscriber" class="Sulu\Bundle\ActivityBundle\Infrastructure\Doctrine\Listener\DomainEventCollectorListener">
             <argument type="service" id="sulu_activity.domain_event_collector"/>
 
-            <tag name="doctrine.event_listener" priority="-256"/>
+            <tag name="doctrine.event_listener" event="onClear" priority="-256"/>
+            <tag name="doctrine.event_listener" event="postFlush" priority="-256"/>
         </service>
 
         <service id="sulu_activity.store_activity_subscriber" class="Sulu\Bundle\ActivityBundle\Application\Subscriber\StoreActivitySubscriber">
